# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title — import paths depend on it
  title: Api
  version: 0.3.0
  description: |
    Painfader DMX Controller API — hardware layer
    3-position lever (−1=NSAR, 0=SCHMERZ, 1=OPIAT)
    
    LED zones (WS2812b via Gledopto 2D-EXMU, Art-Net pixel streaming):
      haube   = Hood matrices (2 × 16×16 = 512 px)
      schmerz = Schmerz band (5 × 16×16 = 1280 px)
      nsar    = NSAR strip (configurable length)
      opiat   = Opiat strip (configurable length)
    
    Fan:   Enttec OpenDMX USB → single DMX channel
    Motor: USB-TTL step/dir board (GRBL or Pololu Tic)
    Lever: GPIO reed contacts (GPI1=N, GPI2=0, GPI3=O)
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dmx
    description: DMX / hardware controller operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dmx/state:
    get:
      operationId: getDmxState
      tags: [dmx]
      summary: Get full controller state
      responses:
        "200":
          description: Current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/hardware-config:
    get:
      operationId: getHardwareConfig
      tags: [dmx]
      summary: Get hardware configuration (IPs, ports, pin numbers…)
      responses:
        "200":
          description: Hardware configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardwareConfig"
    put:
      operationId: updateHardwareConfig
      tags: [dmx]
      summary: Update hardware configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HardwareConfig"
      responses:
        "200":
          description: Updated hardware configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardwareConfig"

  /dmx/mode:
    put:
      operationId: setMode
      tags: [dmx]
      summary: Set operating mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModeInput"
      responses:
        "200":
          description: Mode updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/fan:
    put:
      operationId: setFan
      tags: [dmx]
      summary: Control fan (via OpenDMX USB)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FanInput"
      responses:
        "200":
          description: Fan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/zone/{name}:
    put:
      operationId: setZone
      tags: [dmx]
      summary: Update LED zone pattern (Art-Net pixel)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            enum: [haube, schmerz, nsar, opiat]
          description: "LED zone: haube | schmerz | nsar | opiat"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZonePatternInput"
      responses:
        "200":
          description: Zone updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/motor:
    put:
      operationId: setMotor
      tags: [dmx]
      summary: Control Opiat sign stepper motor (USB-TTL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotorInput"
      responses:
        "200":
          description: Motor command sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/screen:
    get:
      operationId: getScreen
      tags: [dmx]
      summary: Get current screen/video state (media player polls this)
      responses:
        "200":
          description: Screen state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenState"
    put:
      operationId: setScreen
      tags: [dmx]
      summary: Set screen video file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenInput"
      responses:
        "200":
          description: Screen updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/pain-fader:
    put:
      operationId: setPainFader
      tags: [dmx]
      summary: Set fader position (software control / testing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaderPositionInput"
      responses:
        "200":
          description: Position applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/scene:
    put:
      operationId: loadScene
      tags: [dmx]
      summary: Load a named scene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneInput"
      responses:
        "200":
          description: Scene loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/blackout:
    post:
      operationId: blackout
      tags: [dmx]
      summary: Blackout — zero all LED outputs, cut fan
      responses:
        "200":
          description: Blackout applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/hardware-fader:
    post:
      operationId: hardwareFaderInput
      tags: [dmx]
      summary: Receive position from physical 3-position lever (also called by GPIO reader internally)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaderPositionInput"
      responses:
        "200":
          description: Position applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/presets:
    get:
      operationId: getPresets
      tags: [dmx]
      summary: Get all presets
      responses:
        "200":
          description: All presets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresetsState"

  /dmx/presets/{position}:
    put:
      operationId: updatePreset
      tags: [dmx]
      summary: Save preset for position
      parameters:
        - name: position
          in: path
          required: true
          schema:
            type: string
          description: "-1 (NSAR), 0 (SCHMERZ), 1 (OPIAT), or idle"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaderPresetInput"
      responses:
        "200":
          description: Preset saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresetsState"

  /dmx/presets/{position}/capture:
    post:
      operationId: capturePreset
      tags: [dmx]
      summary: Capture current live state as preset
      parameters:
        - name: position
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresetsState"

  /dmx/presets/{position}/apply:
    post:
      operationId: applyPreset
      tags: [dmx]
      summary: Apply a saved preset to live outputs
      parameters:
        - name: position
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preset applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmxState"

  /dmx/preset-timer:
    put:
      operationId: updatePresetTimer
      tags: [dmx]
      summary: Configure idle timer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresetTimerInput"
      responses:
        "200":
          description: Timer config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresetsState"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    # ── Zone pattern ───────────────────────────────────────────────────────────

    RgbColor:
      type: object
      required: [r, g, b]
      properties:
        r: { type: integer, minimum: 0, maximum: 255 }
        g: { type: integer, minimum: 0, maximum: 255 }
        b: { type: integer, minimum: 0, maximum: 255 }

    PatternType:
      type: string
      enum: [solid, pulse, chase, wave, sparkle]

    ZonePattern:
      type: object
      required: [type, primaryColor, secondaryColor, brightness, speed, enabled]
      properties:
        type:
          $ref: "#/components/schemas/PatternType"
        primaryColor:
          $ref: "#/components/schemas/RgbColor"
        secondaryColor:
          $ref: "#/components/schemas/RgbColor"
        brightness:
          type: integer
          minimum: 0
          maximum: 255
        speed:
          type: integer
          minimum: 0
          maximum: 255
          description: "Animation speed: 0=static, 255=~2 Hz cycle"
        enabled:
          type: boolean

    ZonePatternInput:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/PatternType"
        primaryColor:
          $ref: "#/components/schemas/RgbColor"
        secondaryColor:
          $ref: "#/components/schemas/RgbColor"
        brightness:
          type: integer
          minimum: 0
          maximum: 255
        speed:
          type: integer
          minimum: 0
          maximum: 255
        enabled:
          type: boolean

    ZoneState:
      type: object
      required: [pattern, pixelCount]
      properties:
        pattern:
          $ref: "#/components/schemas/ZonePattern"
        pixelCount:
          type: integer
          description: "Physical pixel count for this zone (from hardware config)"

    # ── Full controller state ──────────────────────────────────────────────────

    DmxState:
      type: object
      required: [mode, fan, haube, schmerz, nsar, opiat, motor, screen, painFader, gpio, hardwareConfig, idleTimer]
      properties:
        mode:
          type: string
          enum: [idle, experience]
        fan:
          $ref: "#/components/schemas/FanState"
        haube:
          $ref: "#/components/schemas/ZoneState"
        schmerz:
          $ref: "#/components/schemas/ZoneState"
        nsar:
          $ref: "#/components/schemas/ZoneState"
        opiat:
          $ref: "#/components/schemas/ZoneState"
        motor:
          $ref: "#/components/schemas/MotorState"
        screen:
          $ref: "#/components/schemas/ScreenState"
        painFader:
          $ref: "#/components/schemas/PainFaderState"
        gpio:
          $ref: "#/components/schemas/GpioStatus"
        hardwareConfig:
          $ref: "#/components/schemas/HardwareConfig"
        idleTimer:
          type: object
          required: [enabled, timerSeconds, remaining, triggered]
          properties:
            enabled: { type: boolean }
            timerSeconds: { type: integer }
            remaining: { type: ["number", "null"] }
            triggered: { type: boolean }

    # ── Sub-states ─────────────────────────────────────────────────────────────

    FanState:
      type: object
      required: [speed, enabled, dmxChannel]
      properties:
        speed: { type: integer, minimum: 0, maximum: 255 }
        enabled: { type: boolean }
        dmxChannel: { type: integer }

    MotorState:
      type: object
      required: [position, speed, enabled, simulated]
      properties:
        position:
          type: string
          enum: [up, down, stop]
        speed:
          type: integer
          minimum: 0
        enabled:
          type: boolean
        simulated:
          type: boolean

    ScreenState:
      type: object
      required: [videoFile, enabled]
      properties:
        videoFile: { type: string }
        enabled: { type: boolean }

    PainFaderState:
      type: object
      required: [position, channel]
      properties:
        position:
          type: integer
          minimum: -1
          maximum: 1
        channel:
          type: integer

    GpioStatus:
      type: object
      required: [simulated, position, raw]
      properties:
        simulated: { type: boolean }
        position:
          type: integer
          minimum: -1
          maximum: 1
        raw:
          type: object
          required: [nsar, schmerz, opiat]
          properties:
            nsar:    { type: boolean }
            schmerz: { type: boolean }
            opiat:   { type: boolean }

    # ── Hardware config ────────────────────────────────────────────────────────

    HardwareConfig:
      type: object
      required: [gledopto1, gledopto2, artnetPort, artnetRefreshRate, openDmxPort, fanDmxChannel, motorPort, motorDriverType, motorUpPosition, motorDownPosition, motorMaxSpeed, gpioChip, gpioPinNsar, gpioPinSchmerz, gpioPinOpiat, gpioPollIntervalMs, gpioDebounceMs]
      properties:
        gledopto1:
          type: object
          required: [host, universeStart, haubePixelCount, schmerzPixelCount]
          properties:
            host: { type: string }
            universeStart: { type: integer }
            haubePixelCount: { type: integer }
            schmerzPixelCount: { type: integer }
        gledopto2:
          type: object
          required: [host, universeStart, nsarPixelCount, opiatPixelCount]
          properties:
            host: { type: string }
            universeStart: { type: integer }
            nsarPixelCount: { type: integer }
            opiatPixelCount: { type: integer }
        artnetPort: { type: integer }
        artnetRefreshRate: { type: integer }
        openDmxPort: { type: string }
        fanDmxChannel: { type: integer }
        motorPort: { type: string }
        motorDriverType:
          type: string
          enum: [grbl, tic, simulated]
        motorUpPosition: { type: number }
        motorDownPosition: { type: number }
        motorMaxSpeed: { type: number }
        gpioChip: { type: integer }
        gpioPinNsar: { type: integer }
        gpioPinSchmerz: { type: integer }
        gpioPinOpiat: { type: integer }
        gpioPollIntervalMs: { type: integer }
        gpioDebounceMs: { type: integer }

    # ── Inputs ─────────────────────────────────────────────────────────────────

    ModeInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [idle, experience]

    FanInput:
      type: object
      properties:
        speed: { type: integer, minimum: 0, maximum: 255 }
        enabled: { type: boolean }

    MotorInput:
      type: object
      properties:
        position:
          type: string
          enum: [up, down, stop]
        speed: { type: integer, minimum: 0 }
        enabled: { type: boolean }

    ScreenInput:
      type: object
      properties:
        videoFile: { type: string }
        enabled: { type: boolean }

    FaderPositionInput:
      type: object
      required: [position]
      properties:
        position:
          type: integer
          minimum: -1
          maximum: 1

    SceneInput:
      type: object
      required: [scene]
      properties:
        scene:
          type: string
          enum: [idle, schmerz, opiat, nsar, blackout]

    # ── Presets ────────────────────────────────────────────────────────────────

    FaderPreset:
      type: object
      required: [name, fan, haube, schmerz, nsar, opiat, motor, screen]
      properties:
        name: { type: string }
        fan:
          type: object
          required: [speed, enabled]
          properties:
            speed: { type: integer, minimum: 0, maximum: 255 }
            enabled: { type: boolean }
        haube:
          $ref: "#/components/schemas/ZonePattern"
        schmerz:
          $ref: "#/components/schemas/ZonePattern"
        nsar:
          $ref: "#/components/schemas/ZonePattern"
        opiat:
          $ref: "#/components/schemas/ZonePattern"
        motor:
          type: object
          required: [position, speed, enabled]
          properties:
            position: { type: string, enum: [up, down, stop] }
            speed: { type: integer, minimum: 0 }
            enabled: { type: boolean }
        screen:
          type: object
          required: [videoFile, enabled]
          properties:
            videoFile: { type: string }
            enabled: { type: boolean }

    FaderPresetInput:
      type: object
      properties:
        name: { type: string }
        fan:
          type: object
          properties:
            speed: { type: integer, minimum: 0, maximum: 255 }
            enabled: { type: boolean }
        haube:
          $ref: "#/components/schemas/ZonePatternInput"
        schmerz:
          $ref: "#/components/schemas/ZonePatternInput"
        nsar:
          $ref: "#/components/schemas/ZonePatternInput"
        opiat:
          $ref: "#/components/schemas/ZonePatternInput"
        motor:
          type: object
          properties:
            position: { type: string, enum: [up, down, stop] }
            speed: { type: integer, minimum: 0 }
            enabled: { type: boolean }
        screen:
          type: object
          properties:
            videoFile: { type: string }
            enabled: { type: boolean }

    PresetsState:
      type: object
      required: [positions, idlePreset, idleTimerSeconds, idleTimerEnabled]
      properties:
        positions:
          type: array
          items:
            $ref: "#/components/schemas/FaderPreset"
          description: "[0]=NSAR(-1)  [1]=SCHMERZ(0)  [2]=OPIAT(1)"
        idlePreset:
          $ref: "#/components/schemas/FaderPreset"
        idleTimerSeconds: { type: integer }
        idleTimerEnabled: { type: boolean }

    PresetTimerInput:
      type: object
      properties:
        timerSeconds:
          type: integer
          minimum: 1
          maximum: 3600
        enabled: { type: boolean }