# lib/api-zod/src/generated/types/faderPositionInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Painfader DMX Controller API — hardware layer
 * 3-position lever (−1=NSAR, 0=SCHMERZ, 1=OPIAT)
 *
 * LED zones (WS2812b via Gledopto 2D-EXMU, Art-Net pixel streaming):
 *   haube   = Hood matrices (2 × 16×16 = 512 px)
 *   schmerz = Schmerz band (5 × 16×16 = 1280 px)
 *   nsar    = NSAR strip (configurable length)
 *   opiat   = Opiat strip (configurable length)
 *
 * Fan:   Enttec OpenDMX USB → single DMX channel
 * Motor: USB-TTL step/dir board (GRBL or Pololu Tic)
 * Lever: GPIO reed contacts (GPI1=N, GPI2=0, GPI3=O)
 *
 * OpenAPI spec version: 0.3.0
 */

export interface FaderPositionInput {
  /**
     * @minimum -1
     * @maximum 1
     */
  position: number;
}